M1 RHD additions

Approved-by: Romain Teyssier

+++ feedback/feedback.f90
@@ -63,7 +63,7 @@ subroutine thermal_feedback(s,p,ilevel,msn_loc)
   integer,dimension(1:ndim)::ckey
   integer(kind=8),dimension(0:ndim)::hash_cell
   integer::i,ipart,icellp,ind,idim
-  real(dp)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v
+  real(kind=8)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v
   real(dp)::dx_loc,vol_loc,vol_cell
   real(dp)::mejecta,dloss,dzloss,zloss,ekinetic,ethermal
   real(dp)::birth_time,t_sn,e_sn,dteff,dold
@@ -302,7 +302,7 @@ subroutine mechanical_feedback(s,p,ilevel,msn_loc)
   integer(kind=8),dimension(0:ndim)::hash_cell,hash_nbor
   integer::i,j,k,ipart,icellp,icelln,ind,idim,ivar,ipart_ref
   integer,dimension(1:ndim)::ix
-  real(dp)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v
+  real(kind=8)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v
   real(dp)::dx_loc,vol_loc,vol_cell
   real(dp)::mejecta,ekinetic,ethermal
   real(dp)::birth_time,t_sn,e_sn,dteff,dold,num_SN
@@ -313,6 +313,7 @@ subroutine mechanical_feedback(s,p,ilevel,msn_loc)
   type(oct),pointer::gridp,gridn
   type(msg_large_realdp)::dummy_large_realdp
   logical::ok_level,ok_leaf,ok
+
 
 #ifdef HYDRO
 #if NDIM==3
